folders renamed according to java repo and some code cleanups for the first 4 questions
