+++ Ch 1.Arrays And Strings/Q1_05_One_Away/B.cpp
