+++ Ch 1.Arrays And Strings/Q1_06_String_Compression/6.string_compression.cpp
