+++ Ch 1.Arrays And Strings/Q1_08_Zero_Matrix/8.zero_matrix.cpp
